+++ src/util/fd_1.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   fd_1.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cpak <cpak@student.42seoul.kr>             +#+  +:+       +#+        */
+/*   By: sarchoi <sarchoi@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/27 13:22:24 by cpak              #+#    #+#             */
-/*   Updated: 2022/05/11 00:24:07 by cpak             ###   ########seoul.kr  */
+/*   Updated: 2022/05/11 13:10:45 by sarchoi          ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,15 +56,4 @@ void	set_filein_to_fd(char *path, int fd)
 		return ;
 	}
 	exit (1);
-}
-
-void	set_self_pipe(void)
-{
-	int		fds[2];
-
-	pipe(fds);
-	dup2(fds[0], STDIN_FILENO);
-	dup2(fds[1], STDOUT_FILENO);
-	close(fds[0]);
-	close(fds[1]);
 }
