+++ src/util/fd_3.c
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   fd_3.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sarchoi <sarchoi@student.42seoul.kr>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/11 13:10:49 by sarchoi           #+#    #+#             */
+/*   Updated: 2022/05/11 13:11:05 by sarchoi          ###   ########seoul.kr  */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	set_self_pipe(void)
+{
+	int		fds[2];
+
+	pipe(fds);
+	dup2(fds[0], STDIN_FILENO);
+	dup2(fds[1], STDOUT_FILENO);
+	close(fds[0]);
+	close(fds[1]);
+}
